Convert coadd and zfind scripts to new organization.
